+++ kirc-reactive/src/main/kotlin/de/cmdjulian/kirc/client/ReactiveContainerImageRegistryClient.kt
@@ -4,6 +4,7 @@ import de.cmdjulian.kirc.image.Digest
 import de.cmdjulian.kirc.image.Reference
 import de.cmdjulian.kirc.image.Repository
 import de.cmdjulian.kirc.image.Tag
+
 import de.cmdjulian.kirc.spec.image.ImageConfig
 import de.cmdjulian.kirc.spec.manifest.Manifest
 import de.cmdjulian.kirc.spec.manifest.ManifestList
@@ -89,7 +90,7 @@ interface ReactiveContainerImageRegistryClient {
      *
      * @return the digest of uploaded image
      */
-    fun upload(repository: Repository, reference: Reference, tar: Flux<Byte>): Mono<Digest>
+    fun upload(repository: Repository, reference: Reference, tar: Flux<Byte>, mode: UploadMode = UploadMode.Stream): Mono<Digest>
 
     /**
      * Downloads a docker image for certain [reference].
@@ -141,9 +142,9 @@ fun SuspendingContainerImageRegistryClient.toReactiveClient() = object : Reactiv
         }
     }
 
-    override fun upload(repository: Repository, reference: Reference, tar: Flux<Byte>): Mono<Digest> = mono {
+    override fun upload(repository: Repository, reference: Reference, tar: Flux<Byte>, mode: UploadMode): Mono<Digest> = mono {
         val buffer = Buffer().also { buffer -> tar.collect(buffer::writeByte) }
-        this@toReactiveClient.upload(repository, reference, buffer)
+        this@toReactiveClient.upload(repository, reference, buffer, mode)
     }
 
     override fun download(repository: Repository, reference: Reference): Flux<Byte> = flux {